pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tg-migrate"
version = "0.1.0"
description = "General-purpose Telegram bot token migration CLI tool"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Botable", email = "hello@botable.dev"},
]
keywords = ["telegram", "bot", "migration", "cli"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Communications :: Chat",
]
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "python-telegram-bot>=20.0",
    "rich>=13.0",
]

[project.scripts]
tg-migrate = "tg_migrate.cli:main"

[project.urls]
Homepage = "https://github.com/Botable-dev/tg-migrate"
Repository = "https://github.com/Botable-dev/tg-migrate"
